pyproject.toml
[tool.poetry]
name = "HW_19"
version = "0.1.0"
description = ""
authors = ["Varganova Kate"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
selene = "^2.0.0rc9"
selenium = "^4.23.0"
allure-pytest = "^2.13.5"
appium-python-client = "^4.0.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
